Setup Travis and Rakefile for gem publishing.
Task ci:release will build and push the gem to rubygems.org if and only if it
encounters a version that it's not yet tagged.
Also, it will tinker with gem credentials file in order to allow setting API
key into an env var (which is needed if we want to make pushing the gem happen
in Travis).