Get rid of skeleton repository.
We now build the module from scratch. At this point, it seems to me that adding
the Git stuff makes the project less clear and more complicated in principle:
* what do you do in puppetry and what in the module?
* what if the inherited module (which should have been made customizable at
  some point, since it was the only thing making the design worth it) doesn't
  include folder/files referenced/used in puppetry itself? How to handle the
  possible mismatch?

By not having a half-baked source of data, it looks easier to separate
distinct features like rspec-puppet, rakefile, vagrant, etc. That said, the
design still SUCKS BAD. Needs refactoring + SRP enforcing.

NOTE that now we're a bit back on the features' list, since by cloning a
skeleton repo we had a Rakefile in place, which is not setup yet now.